testutils: move important interfaces to api.go
Release note: None
